Split out VersionsPropsReader, add test
