+++ gradle-baseline-java/src/main/groovy/com/palantir/baseline/plugins/BaselineVersions.java
@@ -16,23 +16,15 @@
 
 package com.palantir.baseline.plugins;
 
-import com.google.common.collect.AbstractIterator;
-import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableMap;
 import java.io.File;
 import java.io.IOException;
 import java.nio.file.Files;
-import java.util.Iterator;
-import java.util.List;
 import java.util.Set;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
 import java.util.stream.Collectors;
-import java.util.stream.Stream;
 import netflix.nebula.dependency.recommender.DependencyRecommendationsPlugin;
 import netflix.nebula.dependency.recommender.RecommendationStrategies;
 import netflix.nebula.dependency.recommender.provider.RecommendationProviderContainer;
-import org.apache.commons.lang3.tuple.Pair;
 import org.gradle.api.Plugin;
 import org.gradle.api.Project;
 import org.gradle.api.artifacts.Configuration;
@@ -63,8 +55,6 @@ import org.gradle.api.plugins.BasePlugin;
  * </pre>
  */
 public final class BaselineVersions implements Plugin<Project> {
-
-    private static final Pattern VERSION_FORCE_REGEX = Pattern.compile("([^:=\\s]+:[^:=\\s]+)\\s*=\\s*([^\\s]+)");
 
     @Override
     public void apply(Project project) {
@@ -107,7 +97,7 @@ public final class BaselineVersions implements Plugin<Project> {
         return file;
     }
 
-    public static Set<String> getResolvedArtifacts(Project project) {
+    static Set<String> getResolvedArtifacts(Project project) {
         return project.getConfigurations().stream()
                 .filter(Configuration::isCanBeResolved)
                 .flatMap(configuration -> {
@@ -126,56 +116,5 @@ public final class BaselineVersions implements Plugin<Project> {
                     }
                 })
         .collect(Collectors.toSet());
-    }
-
-    public static List<Pair<String, String>> readVersionsProps(File propsFile) {
-        ImmutableList.Builder<Pair<String, String>> accumulator = ImmutableList.builder();
-        if (propsFile.exists()) {
-            try (Stream<String> lines = Files.lines(propsFile.toPath())) {
-                new ToggleableIterator(lines.map(String::trim).iterator()).forEachRemaining(line0 -> {
-                    // strip comment
-                    int commentIndex = line0.indexOf("#");
-                    String line = commentIndex >= 0 ? line0.substring(0, commentIndex) : line0;
-                    Matcher matcher = VERSION_FORCE_REGEX.matcher(line);
-                    if (matcher.matches()) {
-                        String propName = matcher.group(1);
-                        String propVersion = matcher.group(2);
-                        accumulator.add(Pair.of(propName, propVersion));
-                    }
-                });
-            } catch (IOException e) {
-                throw new RuntimeException("Error reading " + propsFile.toPath() + " file", e);
-            }
-        } else {
-            throw new RuntimeException("No " + propsFile.toPath() + " file found");
-        }
-        return accumulator.build();
-    }
-
-    private static class ToggleableIterator extends AbstractIterator<String> {
-        private final Iterator<String> iter;
-        private boolean active;
-
-        public ToggleableIterator(Iterator<String> iter) {
-            this.iter = iter;
-            active = true;
-        }
-
-        @Override
-        protected String computeNext() {
-            while (iter.hasNext()) {
-                String line0 = iter.next();
-                if (line0.equals("# linter:ON")) {
-                    active = true;
-                } else if (line0.equals("# linter:OFF")) {
-                    active = false;
-                }
-                if (!active) {
-                    continue;
-                }
-                return line0;
-            }
-            return endOfData();
-        }
     }
 }
